requirements.txt
discord.py[voice]
mutagen